+++ tests/handlers/asgi/test_handle_asgi.py
@@ -1,6 +1,6 @@
 from starlette.status import HTTP_200_OK
 
-from starlite import Controller, MediaType, Response, asgi
+from starlite import Controller, MediaType, Response, ScopeType, asgi
 from starlite.testing import create_test_client
 from starlite.types import Receive, Scope, Send
 
@@ -8,20 +8,20 @@ from starlite.types import Receive, Scope, Send
 def test_handle_asgi() -> None:
     @asgi(path="/")
     async def root_asgi_handler(scope: Scope, receive: Receive, send: Send) -> None:
-        assert scope["type"] == "http"
+        assert scope["type"] == ScopeType.HTTP
         assert scope["method"] == "GET"
         response = Response("Hello World", media_type=MediaType.TEXT, status_code=HTTP_200_OK)
-        await response(scope, receive, send)
+        await response(scope, receive, send)  # type: ignore[arg-type]
 
     class MyController(Controller):
         path = "/asgi"
 
         @asgi()
         async def root_asgi_handler(self, scope: Scope, receive: Receive, send: Send) -> None:
-            assert scope["type"] == "http"
+            assert scope["type"] == ScopeType.HTTP
             assert scope["method"] == "GET"
             response = Response("Hello World", media_type=MediaType.TEXT, status_code=HTTP_200_OK)
-            await response(scope, receive, send)
+            await response(scope, receive, send)  # type: ignore[arg-type]
 
     with create_test_client([root_asgi_handler, MyController]) as client:
         response = client.get("/")